refactor: move function caesar to a separate module
